+++ app/pkg/apierrors/errors_interface.go
@@ -1,11 +1,14 @@
-package api_errors
+package apierrors
 
 type AppError interface {
 	BadRequest() AppError
+	InvalidParameter() ApiError
 	Unauthorized() AppError
 	PaymentRequired() AppError
 	NotFound() AppError
 	TooManyRequests() AppError
 	InternalServerError() AppError
 	ServiceUnavailable() AppError
+
+	New(msg ...string) AppError
 }
